requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
mysql-connector-python==8.1.0
PyMySQL==1.1.0

# Authentication & Security
bcrypt==4.0.1
Werkzeug==2.3.7

# File Upload & Processing
Pillow==10.0.1
python-magic==0.4.27

# Email
Flask-Mail==0.9.1

# Environment Variables
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Date & Time
python-dateutil==2.8.2

# API Documentation
flasgger==0.9.7.1

# Payment Gateways
razorpay==1.3.0
paypalrestsdk==1.13.3

# Shipping
requests==2.31.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Development
flask-shell-ipython==1.4.0

# Utilities
python-slugify==8.0.1
uuid==1.30